+++ examples/timer.rs
@@ -19,6 +19,6 @@ fn main() {
 
     // select all
     enigo.key_down(Key::Control);
-    enigo.key_click(Key::Layout('a'));
+    enigo.key_click(Key::Unicode('a'));
     enigo.key_up(Key::Control);
 }